Add authentication to DELETE methode

+++ controllers/blogs.js
@@ -44,8 +44,23 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (req, res) => {
-  await Blog.findByIdAndDelete(req.params.id);
-  res.status(204).end();
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return res.status(401).json({ error: "token invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+
+  if (!user) {
+    return res.status(400).json({ error: "userId missing or not valid" });
+  }
+  const blog = await Blog.findById(req.params.id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(req.params.id);
+    res.status(204).end();
+  } else {
+    res.status(403).json({ error: "only the creator can delete this blog" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
